Move dropRequestsPerMillion to clusterimpl.go and its test to balancer_test.go
