+++ internal/xds/balancer/clusterimpl/clusterimpl.go
@@ -151,6 +151,19 @@ type clusterImplBalancer struct {
 	lrsReportEndpointMetrics *xdsresource.LRSReportEndpointMetricsConfig // LRS metrics to propagate.
 }
 
+// dropRequestsPerMillion scales a drop overload's numerator and denominator to
+// a number of requests to drop per million. numerator can be as large as a
+// million (a 100% drop), so the multiplication is done in uint64 to avoid
+// overflowing a uint32, and the result is capped at a million because a drop
+// ratio above 100% is treated as 100%.
+func dropRequestsPerMillion(numerator, denominator uint32) uint32 {
+	rpm := uint64(numerator) * million / uint64(denominator)
+	if rpm > million {
+		rpm = million
+	}
+	return uint32(rpm)
+}
+
 // handleClusterConfigLocked updates the internal state of the balancer with the
 // new cluster configuration. It returns true if a new picker needs to be
 // generated as a result of these changes. It must be called with b.mu held.
